[#10] improved svg validation based on marketplace action (#11)
* [#10] Improved SVG validation based on Marketplace Action

Eliminate the installation of xmllint and its usage from the workflow.
Integrate the SVG linter action to validate SVG files in the repository
instead. Adjust the configuration to specify the path for SVG
validation.

Introduce a PowerShell script that formats SVG files with xmlstarlet
and xmllint. Remove trailing whitespace in spaceshipwithexampleheader.
svg to ensure consistency. Update the project tooling for improved SVG
processing.